+++ src/types/OpsLevelData.ts
@@ -55,6 +55,8 @@ export type LevelCheckResults = {
   }
 }
 
+export type CheckResultStatus = 'failed' | 'pending' | 'passed' | 'upcoming_failed' | 'upcoming_pending' | 'upcoming_passed';
+
 export type CheckResult = {
   message: string,
   warnMessage: string | null,
@@ -65,10 +67,17 @@ export type CheckResult = {
     name: string,
     type: string,
     category: {
+      name: string,
+      container: {
+        href: string,
+      },
+    } | null,
+    owner: {
       name: string
+      href: string,
     } | null,
   },
-  status: string
+  status: CheckResultStatus
 }
 
 export interface OpsLevelOverallData {